feat: a research node can sit between two rings
A research's tier is now a distance counted in rings rather than a ring number: a whole number
sits on a ring, anything else between two. Existing whole tiers read back unchanged.
ResearchNetworkPlacement no longer rounds; the editor settles a dropped node instead - onto a ring
within 0.15 of one, otherwise where it was let go, to a tenth of a ring.

+++ Assets/Editor/ResearchTree/ResearchTreeMenu.cs
@@ -60,7 +60,7 @@ namespace Game.EditorTools
             serialized.FindProperty("displayName").stringValue = "Nouvelle recherche";
             serialized.FindProperty("cuCost").floatValue = NewResearchCost;
             serialized.FindProperty("absorptionRatePerSecond").floatValue = NewResearchAbsorption;
-            serialized.FindProperty("tier").intValue = parent != null ? parent.Tier + 1 : 2;
+            serialized.FindProperty("tier").floatValue = parent != null ? parent.Tier + 1f : 2f;
             serialized.FindProperty("angle").floatValue = parent != null ? parent.Angle : 0f;
             if (parent != null)
             {
