+++ Assets/Scripts/Data/ResearchDefinition.cs
@@ -27,8 +27,8 @@ namespace Game.Data
         /// <summary>Ceiling on how many CU per second this research can absorb, even when the reserve holds far more - the runtime rate is min(this, whatever the reserve can currently give).</summary>
         [SerializeField, Min(0f)] float absorptionRatePerSecond;
 
-        /// <summary>Progression tier: the ring this research sits on in the research network (GDD §5.4). With Angle, its whole position there - see ResearchNetworkPlacement.</summary>
-        [SerializeField] int tier;
+        /// <summary>Progression tier: how far from the centre this research sits on the research network, counted in rings (GDD §5.4) - a whole number is on a ring, anything else between two. With Angle, its whole position there - see ResearchNetworkPlacement.</summary>
+        [SerializeField] float tier;
 
         /// <summary>Where on its ring, in degrees counter-clockwise from the right. Placed by hand in the research tree editor; nothing computes or corrects it.</summary>
         [SerializeField] float angle;
@@ -46,7 +46,7 @@ namespace Game.Data
         public IReadOnlyList<ResearchEffect> Effects => effects ?? (IReadOnlyList<ResearchEffect>)System.Array.Empty<ResearchEffect>();
 
         public float AbsorptionRatePerSecond => absorptionRatePerSecond;
-        public int Tier => tier;
+        public float Tier => tier;
         public float Angle => angle;
     }
 }
